# docs/swagger.yaml
basePath: /
definitions:
  models.CreditLine:
    properties:
      allowedRequest:
        type: boolean
      attemptAcceptedNumber:
        type: integer
      attemptNumber:
        type: integer
      cashBalance:
        type: number
      foundingName:
        type: string
      foundingType:
        type: string
      id:
        type: integer
      lastAcceptedRequestDate:
        type: string
      monthlyRevenue:
        type: number
      recommendedCreditLine:
        type: number
      requestedCreditLine:
        type: number
      requestedDate:
        type: string
      requestedServerDate:
        type: string
      state:
        type: string
    type: object
  models.CreditLineRequestBody:
    properties:
      cashBalance:
        type: number
      foundingName:
        type: string
      foundingType:
        type: string
      monthlyRevenue:
        type: number
      requestedCreditLine:
        type: number
      requestedDate:
        type: string
    type: object
  models.CreditLineResponseBody:
    properties:
      cashBalance:
        type: number
      foundingName:
        type: string
      foundingType:
        type: string
      monthlyRevenue:
        type: number
      recommendedCreditLine:
        type: number
      requestedCreditLine:
        type: number
      requestedDate:
        type: string
      requestedServerDate:
        type: string
    type: object
  models.ResponseBody:
    properties:
      data:
        $ref: '#/definitions/models.CreditLineResponseBody'
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server.
  title: Gin Swagger Example API
  version: "1.0"
paths:
  /creditLines:
    get:
      description: get all creditLines
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CreditLine'
            type: array
      summary: Get all creditLines
      tags:
      - creditLine
    post:
      consumes:
      - application/json
      description: Create a creditLine
      parameters:
      - description: creditLine Data
        in: body
        name: creditLine
        required: true
        schema:
          $ref: '#/definitions/models.CreditLineRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseBody'
      summary: Create a creditLine
      tags:
      - creditLine
  /creditLines/{id}:
    get:
      description: Get a creditLine by id
      parameters:
      - description: creditLine ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreditLine'
      summary: Get a creditLine by id
      tags:
      - creditLine
schemes:
- http
swagger: "2.0"